Test default filter
