+++ src/utils/api/deploy/hash-fns.test.js
@@ -1,8 +1,9 @@
 const test = require('ava')
 const hashFns = require('./hash-fns')
+const { defaultFilter } = require('./util')
 
 test('hashes files in a folder', async t => {
-  const { functions, fnShaMap } = await hashFns(__dirname)
+  const { functions, fnShaMap } = await hashFns(__dirname, { filter: defaultFilter })
 
   Object.keys(functions).forEach(path => t.truthy(path, 'each file has a path'))
   t.truthy(fnShaMap, 'fnShaMap is returned')
